requirements.txt
Django==1.6.2
MySQL-python==1.2.5
backports.ssl-match-hostname==3.4.0.2
docker-py==1.1.0
numpy==1.9.2
requests==2.6.0
six==1.9.0
websocket-client==0.26.0
wsgiref==0.1.2
